pyproject.toml
[project]
name = "bg-road-fines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "markdown>=3.8.2",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.0",
]
